requirements.txt
python-decouple
telethon
requests